Loader in scene
